+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 ## Unreleased
 ### Added
+- Parsing of deterministic work metric in nodelog, barrier, and simplex (#27).
 ### Fixed
 ### Changed
 ### Removed
@@ -11,6 +12,7 @@
 - Seed and Time columns in the DataFrame are reported as integer and datetime types, respectively
 - The `get_dataframe(logfiles, timelines=False, prettyparams=False)` API has a new signature
 - A new API `parse(patterns)` is added that returns a `ParseResult()` object with `summary()` and `progress(section)` methods (see README.md for an example usage)
+
 
 ## 1.3.2 - 2022-01-12
 ### Fixed